refactor: 폴더 구조 변경
styled.jsx와 index.jsx 분리 및 공통 컴포넌트(Grid Container, Icon Button) 분리
