+++ src/components/Main/index.jsx
@@ -12,7 +12,7 @@ const Container = styled.main`
   justify-content: center;
   align-items: center;
 
-  max-width: 1320px;
+  width: 100%;
   margin: 140px auto 60px;
 
   height: calc(100vh - 200px);
